+++ backend/handlers/data_helpers.go
@@ -596,69 +596,6 @@ func ReadPostsByGroup(currentUserId int, groupId int) ([]Post, error) {
 	return posts, nil
 }
 
-func InsertGroup(group Group) error {
-	_, err := db.InsertData("groups", group.CreatorProfile.UserId, group.Title, group.Description, time.Now())
-	if err != nil {
-		return errors.New("Error inserting group" + err.Error())
-	}
-	return nil
-}
-func InsertGroupMember(groupId int, userId int) error {
-	_, err := db.InsertData("group_member", groupId, userId)
-	if err != nil {
-		return errors.New("Error inserting group member" + err.Error())
-	}
-	return nil
-}
-func ReadGroup(groupId int) (Group, error) {
-	dbGroups, err := db.FetchData("groups", "groupId", groupId)
-	if err != nil {
-		return Group{}, errors.New("Error fetching group" + err.Error())
-	}
-	if len(dbGroups) == 0 {
-		return Group{}, errors.New("group not found")
-	}
-	dbGroup := dbGroups[0].(db.Group)
-	creator, err := fillSmallProfile(dbGroup.CreatorId)
-	if err != nil {
-		return Group{}, errors.New("Error fetching group creator" + err.Error())
-	}
-	group := Group{
-		GroupId:        dbGroup.GroupId,
-		CreatorProfile: creator,
-		Title:          dbGroup.Title,
-		Description:    dbGroup.Description,
-		Date:           dbGroup.CreationTime,
-	}
-	return group, nil
-}
-func ReadAllGroups(sessionId string) ([]Group, error) {
-	dbGroups, err := db.FetchData("groups", "")
-	if err != nil {
-		return []Group{}, errors.New("Error fetching groups" + err.Error())
-	}
-	if len(dbGroups) == 0 {
-		return []Group{}, errors.New("no group found")
-	}
-	var groups []Group
-	for _, dbGroup := range dbGroups {
-		dbGroup := dbGroup.(db.Group)
-		creator, err := fillSmallProfile(dbGroup.CreatorId)
-		if err != nil {
-			return []Group{}, errors.New("Error fetching group creator" + err.Error())
-		}
-		group := Group{
-			SessionId:      sessionId,
-			GroupId:        dbGroup.GroupId,
-			Title:          dbGroup.Title,
-			Description:    dbGroup.Description,
-			Date:           dbGroup.CreationTime,
-			CreatorProfile: creator,
-		}
-		groups = append(groups, group)
-	}
-	return groups, nil
-}
 func ReadAllUsers(userId int, sessionId string) ([]Profile, error) {
 	dbUsers, err := db.FetchData("users", "")
 	if err != nil {
@@ -703,6 +640,7 @@ func NonMemberUsers(groupId int, userId int, sessionId string) ([]Profile, error
 
 	return nonMembers, nil
 }
+
 func InsertGroupInvitation(senderId int, groupId int, receiverId int, content string) error {
 	_, err := db.InsertData("notifications", receiverId, senderId, groupId, "group_invitation", content, time.Now())
 	if err != nil {
