+++ backend/handlers/groupEvent.go
@@ -1,0 +1,9 @@
+package handlers
+
+/*
+TODO:
+func addEvent(payload json.RawMessage) (Response, error)     {}
+func SendEvents(payload json.RawMessage) (Response, error)   {}
+func EventOptions(payload json.RawMessage) (Response, error) {}
+*/
+// TODO: not sure but maybe we need to add a function to send all users participating in an event
